requirements.txt
requests
python-dotenv
discord.py
python-dateutil
datetime
